+++ Data/Sources/Infrastructure/Audio/AudioRecorderServiceError+Mapping.swift
