--- Data/Sources/Infrastructure/AudioSession/AudioSessionService.swift
@@ -1,26 +1,0 @@
-import AVFoundation
-import Domain
-
-public struct DefaultAudioSessionService: MicrophonePermissionService {
-    public func checkPermission() async -> PermissionStatus {
-        switch AVAudioApplication.shared.recordPermission {
-        case .granted:
-            return .authorized
-        case .denied:
-            return .denied
-        case .undetermined:
-            return .notDetermined
-        @unknown default:
-            return .denied
-        }
-    }
-
-    public func requestPermission() async -> PermissionStatus {
-        let granted = await withCheckedContinuation { continuation in
-            AVAudioApplication.requestRecordPermission { granted in
-                continuation.resume(returning: granted)
-            }
-        }
-        return granted ? .authorized : .denied
-    }
-}
